+++ src/ESP8266_PWM_ISR.hpp
@@ -16,7 +16,7 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.3
+  Version: 1.2.4
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -25,7 +25,8 @@
   1.2.0   K Hoang      29/01/2022 Fix multiple-definitions linker error. Improve accuracy
   1.2.1   K Hoang      30/01/2022 Fix bug. Optimize code
   1.2.2   K Hoang      30/01/2022 DutyCycle to be updated at the end current PWM period
-  1.2.3   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period. Optimize code
+  1.2.3   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period
+  1.2.4   K Hoang      04/03/2022 Fix `DutyCycle` and `New Period` display bugs. Display warning only when debug level > 3
 *****************************************************************************************************************************/
 
 #pragma once
@@ -38,13 +39,13 @@
 #endif
 
 #ifndef ESP8266_PWM_VERSION
-  #define ESP8266_PWM_VERSION           "ESP8266_PWM v1.2.3"
+  #define ESP8266_PWM_VERSION           "ESP8266_PWM v1.2.4"
   
   #define ESP8266_PWM_VERSION_MAJOR     1
   #define ESP8266_PWM_VERSION_MINOR     2
-  #define ESP8266_PWM_VERSION_PATCH     3
+  #define ESP8266_PWM_VERSION_PATCH     4
 
-  #define ESP8266_PWM_VERSION_INT       1002003
+  #define ESP8266_PWM_VERSION_INT       1002004
 #endif
 
 #ifndef TIMER_INTERRUPT_DEBUG
@@ -86,12 +87,19 @@ typedef void (*esp8266_timer_callback)();
 typedef void (*esp8266_timer_callback_p)(void *);
 
 #if !defined(USING_MICROS_RESOLUTION)
-  #warning Not USING_MICROS_RESOLUTION, using millis resolution
+
+  #if (_PWM_LOGLEVEL_ > 3)
+    #warning Not USING_MICROS_RESOLUTION, using millis resolution
+  #endif
+    
   #define USING_MICROS_RESOLUTION       false
 #endif
 
 #if !defined(CHANGING_PWM_END_OF_CYCLE)
-  #warning Using default CHANGING_PWM_END_OF_CYCLE == true
+  #if (_PWM_LOGLEVEL_ > 3)
+    #warning Using default CHANGING_PWM_END_OF_CYCLE == true
+  #endif
+  
   #define CHANGING_PWM_END_OF_CYCLE     true
 #endif